+++ dev/src/Inventory/DataAccessLayer/DAO/PendingDAO.java
@@ -22,7 +22,7 @@ public class PendingDAO extends DalController {
         pendingIdentityMap = new PendingIdentityMap();
     }
 
-    public void InsertPending(String name, String producer, int price, int quantity) {
+    public void InsertPending(String name, String producer, double price, int quantity) {
         String sql = "INSERT INTO Pending(name,producer," +
                 "price,quantity) " +
                 "VALUES(?,?,?,?)";
@@ -37,7 +37,7 @@ public class PendingDAO extends DalController {
 
             pstmt.executeUpdate();
 
-            pendingIdentityMap.addPending(new Pair<String,String>(name,producer),new Pair<Integer,Integer>(price,quantity));
+            pendingIdentityMap.addPending(new Pair<String,String>(name,producer),new Pair<Double,Integer>(price,quantity));
             //return productIdentityMap.addProduct(new Product(Id,name,producer,buyingPrice,sellingPrice,categories,minQuantity));
         } catch (SQLException e) {
             System.out.println(e.getMessage());
@@ -45,7 +45,7 @@ public class PendingDAO extends DalController {
         }
     }
 
-    public Map<Pair<String,String>,Pair<Integer,Integer>> SelectAll() {
+    public Map<Pair<String,String>,Pair<Double,Integer>> SelectAll() {
         if(pendingIdentityMap.isPulled_all_data())
             return pendingIdentityMap.getPendings();
         String sql = "SELECT * FROM Pending";
@@ -59,7 +59,7 @@ public class PendingDAO extends DalController {
                 String producer = rs.getString("producer");
                 if(!pendingIdentityMap.pendingExists(name,producer))
                     pendingIdentityMap.addPending(new Pair<>(name,producer),
-                            new Pair<>(rs.getInt("price"),rs.getInt("quantity") ));
+                            new Pair<>(rs.getDouble("price"),rs.getInt("quantity") ));
             }
             pendingIdentityMap.setPulled_all_data(true);
             return pendingIdentityMap.getPendings();
